+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/project/renderer/commonmark/CommonMarkChangelogRendererTest.java
@@ -11,21 +11,21 @@ import static net.splitcells.dem.testing.TestTypes.INTEGRATION_TEST;
 import static net.splitcells.website.server.project.ProjectRenderer.projectRenderer;
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class CommonMarkReadmeExtensionTest {
+public class CommonMarkChangelogRendererTest {
     /**
      * TODO Test extension explicitly.
      */
     @Tag(INTEGRATION_TEST)
     @Test
-    public void testReadMeLayout() {
+    public void testLayout() {
         final var testSubject = projectRenderer
                 ("public"
-                        , Path.of("../net.splitcells.dem")
+                        , Path.of("../..")
                         , Path.of("../net.splitcells.website.default.content/src/main/xsl/net/splitcells/website/den/translation/to/html/")
                         , Path.of("net.splitcells.website.default.content/src/main/resources/html")
-                        , "/net/splitcells/dem"
+                        , "/net/splitcells/network"
                         , a -> Optional.empty()
                         , Config.create());
-        assertThat(testSubject.projectPaths()).contains(Path.of("net/splitcells/dem/README.html"));
+        assertThat(testSubject.projectPaths()).contains(Path.of("net/splitcells/network/CHANGELOG.html"));
     }
 }